Fairly large changes
user now adds name of standard during prompts, and is used to create folders with name. If custom directory is selected, it checks if it matches the given standard name. New functions such as names.concatDirectories have been added for more reliablility, and the ordering of the items to download has been changed